chore: migration WIP - single error for now

+++ gae_app/debtstracker/models/transfer.go
@@ -241,8 +241,8 @@ type TransferData struct {
 	IsOutstanding bool
 	Currency      money.Currency // Should be indexed for loading outstanding transfers
 	//
-	ReceiptsSentCount int64   `datastore:",noindex,omitempty"`
-	ReceiptIDs        []int64 `datastore:",noindex"`
+	ReceiptsSentCount int   `datastore:",noindex,omitempty"`
+	ReceiptIDs        []int `datastore:",noindex"`
 }
 
 // AmountReturned returns amount returned to counterparty
